Move ConfigStoreBasedTemplateCache to tenant space
