+++ components/template-mgt/org.wso2.carbon.identity.template.mgt/src/main/java/org/wso2/carbon/identity/template/mgt/cache/ConfigStoreBasedTemplateCacheEntry.java
@@ -18,7 +18,7 @@
 
 package org.wso2.carbon.identity.template.mgt.cache;
 
-import org.wso2.carbon.identity.application.common.cache.CacheEntry;
+import org.wso2.carbon.identity.core.cache.CacheEntry;
 import org.wso2.carbon.identity.template.mgt.model.Template;
 
 /**
